Update START_PRINT_macro
Added replacement Quad Gantry Leveling that should simply use all slicer settings when sent

also, no need to heat the bed for a QGL... if the bed is flush, or even taco'd, the gantry will still touch all 4 corners to flush itself out. No need to heat up the bed for this.